my_cdk/requirements.txt
aws-cdk-lib
constructs
jsii
